feat(table/date-cell-popover): update remind selection
