+++ src/components/common/Menu/index.js
@@ -1,5 +1,5 @@
 import { Pressable } from '@react-aria/interactions'
-import React from 'react'
+import { CheckIcon } from '@/components/icons'
 
 const MenuItem = ({ item, isActive, onChange }) => {
   const { value, label } = item || {}
@@ -7,6 +7,7 @@ const MenuItem = ({ item, isActive, onChange }) => {
     <Pressable onPress={() => onChange(value)}>
       <div className="flex-between h-7 cursor-pointer rounded px-2 transition-colors hover:bg-[rgba(55,_53,_47_,0.08)]">
         <span className="text-sm">{label}</span>
+        {isActive && <CheckIcon className="size-3" />}
       </div>
     </Pressable>
   )
@@ -14,7 +15,6 @@ const MenuItem = ({ item, isActive, onChange }) => {
 
 const Menu = ({ list, value, onChange }) => {
   const renderItem = (item) => {
-    const { value } = item || {}
     const isActive = value === item.value
     return <MenuItem key={item.value} isActive={isActive} item={item} onChange={onChange} />
   }
