+++ src/constants/texts.js
@@ -112,9 +112,9 @@ export const DATE_FORMAT_TEXT = {
 export const NOT_INCLUDE_TIME_REMIND_TEXTS = {
   [NOT_INCLUDE_TIME_REMIND_TYPES.NONE]: 'None',
   [NOT_INCLUDE_TIME_REMIND_TYPES.ON_DAY_OF_EVENT]: 'On day of event',
-  [NOT_INCLUDE_TIME_REMIND_TYPES.ONE_DAY_BEFORE]: 'One day before',
-  [NOT_INCLUDE_TIME_REMIND_TYPES.TWO_DAYS_BEFORE]: 'Two days before',
-  [NOT_INCLUDE_TIME_REMIND_TYPES.ONE_WEEK_BEFORE]: 'One week before',
+  [NOT_INCLUDE_TIME_REMIND_TYPES.ONE_DAY_BEFORE]: 'One day before (9:00 AM)',
+  [NOT_INCLUDE_TIME_REMIND_TYPES.TWO_DAYS_BEFORE]: 'Two days before (9:00 AM)',
+  [NOT_INCLUDE_TIME_REMIND_TYPES.ONE_WEEK_BEFORE]: 'One week before (9:00 AM)',
 }
 
 export const INCLUDE_TIME_REMIND_TEXTS = {
@@ -122,7 +122,7 @@ export const INCLUDE_TIME_REMIND_TEXTS = {
   [INCLUDE_TIME_REMIND_TYPES.AT_TIME_OF_EVENT]: 'At time of event',
   [INCLUDE_TIME_REMIND_TYPES.FIVE_MINUTES_BEFORE]: 'Five minutes before',
   [INCLUDE_TIME_REMIND_TYPES.TEN_MINUTES_BEFORE]: 'Ten minutes before',
-  [INCLUDE_TIME_REMIND_TYPES.FIFTEEN_MINUTES_BEFORE]: 'Five minutes before',
+  [INCLUDE_TIME_REMIND_TYPES.FIFTEEN_MINUTES_BEFORE]: 'Fifteen minutes before',
   [INCLUDE_TIME_REMIND_TYPES.THIRTY_MINUTES_BEFORE]: 'Thirty minutes before',
   [INCLUDE_TIME_REMIND_TYPES.ONE_HOUR_BEFORE]: 'One hour before',
   [INCLUDE_TIME_REMIND_TYPES.TWO_HOURS_BEFORE]: 'Two hours before',
